Ändrade int till float

+++ miniräknare.py
@@ -3,11 +3,11 @@ __author__ = 'Andreas'
 val = input("Vad vill du göra? addition, subtraktion, multiplication eller division?")
 
 if val == ("division"):
-    siffra_1 = int(input("Skriv in tal 1:"))
-    siffra_2 = int(input("Skriv in tal 2:"))
+    siffra_1 = float(input("Skriv in tal 1:"))
+    siffra_2 = float(input("Skriv in tal 2:"))
     print("Ditt svar är:", siffra_1 / siffra_2)
 
 if val == ("addition"):
-    siffra_1 = int(input("Skriv in tal 1:"))
-    siffra_2 = int(input("Skriv in tal 2:"))
+    siffra_1 = float(input("Skriv in tal 1:"))
+    siffra_2 = float(input("Skriv in tal 2:"))
     print("Ditt svar är:", siffra_1 + siffra_2)
